prevent executing initializing process multiple times

+++ gen_config.py
@@ -1,16 +1,19 @@
 import yaml 
 import os
-os.makedirs("/app/app_config",exist_ok=True)
-os.makedirs("/app/db",exist_ok=True)
-defaultData = dict(
-    db = dict(
-        driver = '',
-        connection = ''
-    ),
-    node_address ='http://localhost',
-    port='8090',
-    public_registration= True,
-    post_registry_url= "/post-regist.html"
-)
-with open('/app/app_config/config.yaml', 'w') as outfile:
-    yaml.dump(defaultData, outfile, default_flow_style=False)
+from pathlib import Path
+
+if not Path('/app/app_config/config.yaml').exists():
+   os.makedirs("/app/app_config",exist_ok=True)
+   os.makedirs("/app/db",exist_ok=True)
+   defaultData = dict(
+      db = dict(
+         driver = '',
+         connection = ''
+      ),
+   node_address ='http://localhost',
+   port='8090',
+   public_registration= True,
+   post_registry_url= "/post-regist.html"
+   )
+   with open('/app/app_config/config.yaml', 'w') as outfile:
+      yaml.dump(defaultData, outfile, default_flow_style=False)
